+++ evac/worker.py
@@ -35,131 +35,6 @@ if 'AAMKS_SKIP_CFAST' in os.environ:
         SIMULATION_TYPE = 'NO_CFAST'
 
 
-'''Designed to be used as a  parameter of EvacEnv class from evac.rvo2_dto'''
-class Detection:
-    def __init__(self, vars_conf: dict, floor: EvacEnv):
-        self.project_conf = vars_conf    # more data in conf from worker vars to be inherited in the future
-        self.floor = floor
-        self.time = floor.current_time
-        self.config = floor.config
-        self.conditions = floor.smoke_query.compa_conditions
-        self.rooms, self.sensors = self._get_rooms_sensors(floor.smoke_query.all_compas)
-        self.room_heights = {room: None for room in self.rooms}
-        self.state = dict(rooms={room: None for room in self.rooms}, floor=None)
-
-    @staticmethod
-    def _get_rooms_sensors(all_compas):
-        rooms, sensors = [], []
-        for entity in all_compas:
-            if entity.startswith(('sd', 'sp', 'hd')):
-                sensors.append(entity)
-            elif entity.startswith('t_'):
-                # we don't care about targets - those are for fire spread
-                continue
-            else:
-                rooms.append(entity)
-        return rooms, sensors
-
-    def _initialize(self):
-        # initial heights
-        initial = copy.deepcopy(self.conditions)
-        for r in self.rooms:
-            self.room_heights[r] = initial[r]['HGT']
-        del initial
-
-        # set fire origin room detection to 0.001 (can't be 0 because bool(0) = False)
-        self.state['rooms'][self.project_conf['FIRE_ORIGIN']] = .001
-
-    def _od_to_vis(self, optical_density):
-        # convert optical density to visibility
-        if optical_density:
-            return min([30, self.project_conf['c_const'] / (optical_density * log(10))])
-        else:
-            return 30
-
-    def _is_fire_from_sensor(self, sensor):
-        # return sensor state
-        return bool(self.conditions[sensor]['SENSACT'])
-
-    def _is_fire_symptom(self, room: str):
-        actual_ulod = self.conditions[room]['ULOD']
-        actual_height = self.conditions[room]['HGT']
-        initial_height = self.room_heights[room]
-        # check for fire symptoms in room
-        if not initial_height and not actual_height:
-            # one-zone model
-            if self._od_to_vis(actual_ulod) >= self.config['LOWEST_VIS']:
-                return True
-        elif actual_height <= self.config['PRE_EVAC_TIME_ZONE_REDUCTION'] * initial_height:
-            # two-zone model
-            if self._od_to_vis(actual_ulod) >= self.config['LOWEST_VIS']:
-                return True
-        return False
-
-    def _update_floor_state(self):
-        # iterate over sensors to evaluate their state
-        if not self.state['floor']:
-            for sensor in self.sensors:
-                if self._is_fire_from_sensor(sensor):
-                    self.state['floor'] = self.time
-
-    def _update_rooms_state(self):
-        # iterate over rooms to evaluate rooms' conditions and state
-        for room in self.rooms:
-            if self.state['rooms'][room]:
-                continue
-            if self._is_fire_symptom(room):
-                self.state['rooms'][room] = self.time
-
-    def _delay_from_room(self, room: str, pre_evac: float):
-        # calculate total delay time for room that is in fire
-        room_detection = self.state['rooms'][room]
-        if room_detection:
-            return room_detection + pre_evac
-        else:
-            return self.config['DETECTION_TIME']
-
-    def _delay_from_floor(self, alarm: float, pre_evac: float):
-        # calculate total delay time for room that is in fire
-        floor_detection = self.state['floor']
-        if floor_detection:
-            return floor_detection + alarm + pre_evac
-        else:
-            return self.config['DETECTION_TIME']
-
-    def _get_pedestrian_delay(self, evacuee: Evacuee):
-        floor_data = self.project_conf['FLOORS_DATA'][str(self.floor.floor)]
-
-        room_delay = self._delay_from_room(evacuee.detection_compa, evacuee.detection_constituents['pre_evac_fire_origin'])
-        if evacuee.detection_compa == self.project_conf['FIRE_ORIGIN']:
-            return room_delay
-        else:
-            floor_delay = self._delay_from_floor(floor_data['ALARMING'], evacuee.detection_constituents['pre_evac'])
-            return min(floor_delay, room_delay)
-
-    def _update_delays(self):
-        # iterate over evacuees and if still not moving set them proper delay
-        for evacuee in self.floor.evacuees.pedestrians:
-            if evacuee.velocity != (0, 0):
-                continue
-            new_delay = self._get_pedestrian_delay(evacuee)
-            if evacuee.pre_evacuation_time > new_delay:
-                evacuee.pre_evacuation_time = new_delay
-
-    def update(self):
-        self.time = round(self.floor.current_time, 2)
-        if self.time == 0:
-            self._initialize()
-        else:
-            # update state variables if needed
-            self._update_floor_state()
-            self._update_rooms_state()
-            # update evacuees pre-evac times
-            self._update_delays()
-        # return floor detection time
-        return self.state['floor']
-
-
 class Worker:
 
     def __init__(self, redis_worker_pwd = None, AA=None):
@@ -697,8 +572,6 @@ class Worker:
             shutil.rmtree(self.working_dir + "/door_opening_changes")
         os.mkdir(self.working_dir + "/door_opening_changes")
         self.initialize_doors_opening_level()
-        for floor in self.floors:
-            floor.detection = Detection(self.vars['conf'], floor)
 
         while 1:
             time_frame += cfast_step    # increase upper limit of time_frame
